🎨 Refactor code structure: separate ML and DB logic
- Moved ML server communication to `ml_client.py`
- Moved database communication to `db_client.py`
- Updated `app.py` to use the new module structure
