+++ food_classifier/src/service_ui/app.py
@@ -1,39 +1,10 @@
 import gradio as gr
 import cv2
 import numpy as np
-from unittest.mock import patch # TODO(GideokKim): Remove this import when ML server is ready
-import requests
 import json
-
-def get_food_info_from_db(food_name):
-    """
-    Query the nutrition database for food information based on the food name.
-    TODO(GideokKim): Connect to Azure Database for MySQL server
-    """
-    # This function will be mocked in tests
-    pass
-
-def get_food_prediction_from_ml_server(img_bytes):
-    """
-    Send image to ML server and get food prediction.
-    TODO(GideokKim): Connect to Custom Vision Server
-    """
-
-    # TODO(GideokKim): Uncomment when ML server is ready
-    # # Send image to ML server
-    # response = requests.post(
-    #     "http://localhost:8000/predict",  # ML server endpoint
-    #     files={"file": img_bytes},
-    #     timeout=30
-    # )
-    
-    # if response.status_code == 200:
-    #     result = response.json()
-    #     food_name = result.get("food_name", "Unknown")
-    #     confidence = result.get("confidence", 0.0)
-
-    # This function will be mocked in tests
-    pass
+from unittest.mock import patch # TODO(GideokKim): Remove this import when ML server is ready
+from ml_client import get_food_prediction_from_ml_server
+from db_client import get_food_info_from_db
 
 def process_image(image):
     """
@@ -100,7 +71,7 @@ demo = gr.Interface(
 # Run server
 if __name__ == "__main__":
     # Mock the database and ML server functions for testing
-    with patch('__main__.get_food_info_from_db', return_value={
+    with patch('db_client.get_food_info_from_db', return_value={
         "food_name": "김치찌개",
         "serving_size": "1인분 (300g)",
         "nutrition": {
@@ -111,7 +82,7 @@ if __name__ == "__main__":
             "sodium": "1500mg",
             "sugar": "3g"
         }
-    }), patch('__main__.get_food_prediction_from_ml_server', return_value=("김치찌개", 95.7)):
+    }), patch('ml_client.get_food_prediction_from_ml_server', return_value=("김치찌개", 95.7)):
         demo.launch(
             server_name="0.0.0.0",  # Allow external connections
             server_port=7860,       # Specify port
