+++ food_classifier/src/service_ui/db_client.py
@@ -1,0 +1,7 @@
+def get_food_info_from_db(food_name):
+    """
+    Query the nutrition database for food information based on the food name.
+    TODO(GideokKim): Connect to Azure Database for MySQL server
+    """
+    # This function will be mocked in tests
+    return None
